# app-service/src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Product API
  version: 1.0.0
paths:
  /products:
    get:
      summary: Get all products
      operationId: getAllProducts
      tags:
        - Product
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DataProductListDto'
    post:
      summary: Create product
      operationId: createProduct
      tags:
        - Product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Invalid input data
        '409':
          description: Email already exists

  /products/{productId}:
    get:
      summary: Get product by ID
      operationId: getProductById
      tags:
        - Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Found product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found

    put:
      summary: Update product
      operationId: updateProduct
      tags:
        - Product
      parameters:
        - in: path
          name: productId
          description: ID of the product to update
          schema:
            type: long
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'

    delete:
      summary: Delete product by ID
      operationId: deleteProduct
      tags:
        - Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found

components:
  schemas:
    ProductDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Laptop"
        price:
          type: number
          example: 1299.99
      required:
        - name
        - price


    DataProductListDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductDto'